Final fixups to get disabling of exports via showmount.
I'm still learning the layout of both nfs as well as it's common-graphics
code, which is also new to me.  After building things and testing I found
that the DUMP command wasn't the one called during a
but it was EXPORT instead.  I modified the code to do things correctly.  Then
I found that the configuration writing/reading code needed to be updated as
well to write out a config for the *showmount-disabled* variable.  Now
everything is displaying well and functioning.  I added some debugging stuff
to show when *mountd-debug* is enabled, so people can tell when an EXPORT is
disabled.  Also added some code to notify when reloading the config, which I
may remove in the future.

I'll need to do some more testing and ensure that the EXPORT command isn't
used in some other way that I'm unaware of to ensure that disabiling it's
functionality isn't crippling the product in any way.