Adding a check that we don't recurse forever when finding exports.
Maybe _the_ cause of the 100% cpu bug, but definiately one of the of
them.  From talking with Duane it's possible for locate-nearest-export
to have checked *exports* and found it had an export, then when it
calls locate-nearest-export-1 to have a change of context to another
process that might change *exports* (like re-reading the
configuration) and then when locate-nearest-export-1 continues to have
it get into a state where it recurses on (locate-nearest-export-1 "/c"
1), which will never return if *exports* is nil.

This change adds a check to see if *exports* is nil in
locate-nearest-export-1, if so then it exits.  Duane also suggested
passing the value of *exports* from locate-nearest-export on to the
-1 and -2 function calls to ensure that the value doesn't change, but
in some ways if the value does change that could be the more desired
behavior.