Only cache file attributes for a max of *attr-cache-reap-time* seconds.
Prior to this commit, the expiration for cached file attributes would be
extended each time they were accessed.  This could be a problem in
the following scenario:

* An NFS client is repeatedly stat'ing a file to see if its attributes had changed.

* The attributes of the file were altered outside of Allegro NFS's knowledge

In this case the new file attributes would never be returned... at least not
until the client stopped probing for a sufficiently long time.