Improve timing precision of hammernfs
Prior to this commit, hammernfs used time() to determine if the test
interval had completed.  time() only has one second resolution.
Changed to use gettimeofday which has (potentially) microsecond
resolution.  Actual resolution will depend on the OS but it'll surely
be better than one second.

Added the number of reads to the report.