Further mac os x build adjustments
Got rid of all warnings when building hammernfs on Mac OS X.

Also:

Makefile:
 Clean hammernfs in 'clean' target.
 Added 'tags' target.

hammernfs.c: Use strerror() to expand error codes in lookup2 and
lookup3.  This assumes that the returned code is the same as the Unix
errno, which is generally true.  One of these I'll need to make a
mapping function.