rfe14498: handle tcp connection reset errors
The new test/test-conn-reset.c can be used to demonstrate the problem
from spr43071, where the service would die because of the error.

nfsd uses the new function sunrpc:rpc-receive-and-handle-message,
which does the work of catching the error and removing the stream from
the list of client sockets in the server.

Reduced redundancy in Makefile.

Delete test-conn-test in the 'clean' target.

Remove unused nfs_read function in test-conn-reset.c

Change-Id: I2a70058a00cf477537adad3011aef1e0aef17976
Reviewed-on: https://gerrit.franz.com:9080/6136
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>