6.2.beta.1: return directory entries in OS supplied order
Added (nreverse res) at the end of the unicode-directory function so
that filenames will be returned in the order that the OS returned
them.  This means that . and .. should be first (although I found that
. and .. are entirely missing when getting a directory listing of a
root directory).

Also removed an old #+ignored augmented-directory function.

Added release note item to README.md.

Change-Id: I739c4fc8c0ddb638cf30fc289d3042a95c2fdece
Reviewed-on: https://gerrit.franz.com:9080/6543
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Tested-by: Kevin Layer <layer@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>