bug24246: Check for and invalidate out of date file handles
I made changes to the function (insert-fhandle) which handles
insertion of new file handles into caching data structures.
Previously the function assumed that the id of a new file handle had
never been used before.  However, if a file or directory has been
renamed outside of the knowledge of Allegro NFS, then it's possible
that the "new" file/dir will have the same file id as an existing
entry.

I changed the function to check for the existence of a file handle w/
the same id before doing anything else [Note: It only performs this
check if the current operation is not an internal rename operation].
If such a file handle is found, it is invalidated before continuing w/
the insertion of the new entry.  This prevents Allegro NFS from
holding on to out-of-date information.

The new check for existing file ids will cost additional CPU time when
new file handles are being inserted.  New file handles are inserted
the first time a file is "seen" by Allegro NFS.  The check amounts to
a hash table lookup so the cost is low.

Simplified the way that export name canonicalization and searching
works.  Added some related utility functions.

Made locate-nearest-export-by-nfs-path operate correctly when using
submounts.

Change-Id: I219295283c3bd2ad6e64c527655bc6de9d798494
Reviewed-on: https://gerrit.franz.com:9080/6585
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>