Bugfix for prior commits
Fixed recently-introduced bug in insert-fhandle when aNFS is renaming
a file on behalf of a client.

Sprinkled comments and code documentation here and there.

Simplified the way that export name canonicalization works.  Added
some related utility functions.

Made locate-nearest-export-by-nfs-path operate correctly when using
submounts.

This commit to be squashed later.