rfe14956: Make Allegro NFS resilient to nsm-state file corruption
Bumped version to 6.3.beta.2.

For whatever reason, every once in a while a user reports a problem
restarting Allegro NFS due to a correct nsm-state file (often filled
with nulls).  It's unclear why this happens but I did find one
candidate: The fsync call in nsm-state-save was not preceded by a
finish-output, so it's possible that nothing was being fsync'd.  I
took care of that.  I also updated the state file reading code to bail
out and revert to default initial state if there's a problem reading
the file.  The broken file will be renamed to nsm-state.broken for
examination.

Other changes:

Updated my-rename to convert pathnames to namestrings as needed.

hexdump-file function added to utils.cl.

Change-Id: I63ecfa6e2d72cb3bff3baf61754fb4870cd51015
Reviewed-on: https://gerrit.franz.com:9080/7622
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>