rfe15068: tell user when reboot is needed
and refuse to install if "file renames on reboot" are pending.

Bump version to 6.3.0.  This will be release candidate 1.

Also, remove a statement in the uninstall script that caused the
reboot flag to be always set when nfs.cfg was preserved.  Basically,
we were doing

  RMDir /rebootok "$INSTDIR"

when nfs.cfg was preserved, and this caused the above to fail, which
set the reboot flag (because /rebootok was given).  Now, when nfs.cfg
is preserved, we don't execute the above statement.

Make sure all MessageBox's before aborts consistently use MB_ICONSTOP.

Use a single case for "RMDir".

Also, tested and built with NSIS 3.01, which is now installed on thunder.

Change-Id: I4d250de2c5800b0214689ba093b38a9745c2cc5c
Reviewed-on: https://gerrit.franz.com:9080/7936
Tested-by: Kevin Layer <layer@franz.com>
Reviewed-by: Ahmon Dancy <dancy@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>