rfe15679: no caching open files if *open-file-reap-time* = 0
6.4.beta.4

When *open-file-reap-time* (it used to be called *openfilereaptime*) is 0,
don't cache open files in the *open-file-cache* hash table.  This will
incur a performance penalty, but one that some customers may be willing
to bear for the benefit of not running into sharing violations on
Windows, when accessing files modified or created by clients.

Change-Id: I335aa7c365705c289b80e19bf8e1bd389f07f4ab
Reviewed-on: https://gerrit.franz.com:9080/c/nfs/+/11098
Reviewed-by: Martin Mikelsons <mm@franz.com>
Reviewed-by: Kevin Layer <layer@franz.com>
Tested-by: Kevin Layer <layer@franz.com>